componentized the date input field in adding object
